% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dumpGEM}
\alias{dumpGEM}
\title{Dumps all GEM data}
\usage{
dumpGEM(keep_zip = FALSE, overwrite = FALSE)
}
\arguments{
\item{keep_zip}{if TRUE, leave the downloaded zip files in your 'gem.path',
if FALSE, they will be deleted, default FALSE}

\item{overwrite}{if TRUE, skips downloading already existing data files, if
FALSE, they will be downloaded again and overwritten, default FALSE}
}
\description{
Downloads all public GEM data from the GEM project website
}
\examples{
\dontrun{
dumpGEM()
}
}

