% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_gem_data}
\alias{get_gem_data}
\title{Downloads a set of GEM data and restores it in gem.path}
\usage{
get_gem_data(gem.file.id, overwrite = FALSE, keepZip = FALSE)
}
\arguments{
\item{gem.file.id}{if true, leave the downloaded zip files in your
'gem.path', if FALSE, they will be deleted}

\item{keepZip}{if true, leave the downloaded zip files in your 'gem.path',
if FALSE, they will be deleted}

\item{overwirte}{if TRUE, skips downloading already existing data files, if
FALSE, they will be downloaded again and owerwritten, default FALSE}
}
\description{
Downloads a set of public GEM data from the GEM project website
}

