% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGEMData}
\alias{getGEMData}
\title{Downloads a set of GEM data and restores it in gem.path}
\usage{
getGEMData(gem_file_id, overwrite = FALSE, keep_zip = FALSE)
}
\arguments{
\item{gem_file_id}{if true, leave the downloaded zip files in your
'rgem.data.dir', if FALSE, they will be deleted}

\item{keep_zip}{if true, leave the downloaded zip files in your 'gem.path',
if FALSE, they will be deleted}

\item{overwirte}{if TRUE, skips downloading already existing data files, if
FALSE, they will be downloaded again and owerwritten, default FALSE}
}
\description{
Downloads a set of public GEM data from the GEM project website
}

