% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dump_gem}
\alias{dump_gem}
\title{Dumps all GEM data}
\usage{
dump_gem(keepZip = FALSE)
}
\arguments{
\item{keepZip}{if true, leave the downloaded zip files in your 'gem.path', if FALSE, they will be deleted, default FALSE}
}
\description{
Downloads all public GEM data from the GEM project website
}
\examples{
dump_gem()
}

